+++ adodb-csvlib.inc.php
@@ -29,11 +29,13 @@ global $ADODB_INCLUDED_CSV;
 $ADODB_INCLUDED_CSV = 1;
 
 	/**
- 	 * convert a recordset into special format
+ 	 * Convert a recordset into special format
 	 *
-	 * @param rs	the recordset
+	 * @param ADORecordSet  $rs the recordset
+	 * @param ADOConnection $conn
+	 * @param string        $sql
 	 *
-	 * @return	the CSV formatted data
+	 * @return string the CSV formatted data
 	 */
 	function _rs2serialize(&$rs,$conn=false,$sql='')
 	{
@@ -74,7 +76,7 @@ $ADODB_INCLUDED_CSV = 1;
 
 		$savefetch = isset($rs->adodbFetchMode) ? $rs->adodbFetchMode : $rs->fetchMode;
 		$class = $rs->connection->arrayClass;
-		$rs2 = new $class();
+		$rs2 = new $class(-1); // Dummy query Id
 		$rs2->timeCreated = $rs->timeCreated; # memcache fix
 		$rs2->sql = $rs->sql;
 		$rs2->oldProvider = $rs->dataProvider;