+++ adodb-xmlschema.inc.php
@@ -1206,7 +1206,7 @@ class dbQuerySet extends dbObject {
  * @tutorial getting_started.pkg
  *
  * @author Richard Tango-Lowy & Dan Cech
- * @version $Revision: 1.12 $
+ * @version 1.12
  *
  * @package axmls
  */