+++ adodb-xmlschema03.inc.php
@@ -585,8 +585,11 @@ class dbTable extends dbObject {
 					// Option has an argument.
 					if( is_array( $opt ) ) {
 						$key = key( $opt );
-						$value = $opt[key( $opt )];
-						@$fldarray[$field_id][$key] .= $value;
+						$value = $opt[$key];
+						if(!isset($fldarray[$field_id][$key])) {
+							$fldarray[$field_id][$key] = "";
+						}
+						$fldarray[$field_id][$key] .= $value;
 					// Option doesn't have arguments
 					} else {
 						$fldarray[$field_id][$opt] = $opt;
@@ -1317,7 +1320,7 @@ class dbQuerySet extends dbObject {
  * @tutorial getting_started.pkg
  *
  * @author Richard Tango-Lowy & Dan Cech
- * @version $Revision: 1.62 $
+ * @version 1.62
  *
  * @package axmls
  */